- refactor: reorganising files in msg handling module
- refactor: moving msg_filter mod into core mod
- refactor: moving core module out of routing_api module
