+++ src/routing/routing_api/tests/mod.rs
@@ -18,6 +18,7 @@ use crate::messaging::{
     DstInfo, DstLocation, MessageType, SectionAuthorityProvider, SrcLocation,
 };
 use crate::routing::{
+    core::{RESOURCE_PROOF_DATA_SIZE, RESOURCE_PROOF_DIFFICULTY},
     dkg::{
         test_utils::{prove, section_signed},
         ProposalUtils,
@@ -29,7 +30,6 @@ use crate::routing::{
     node::Node,
     peer::PeerUtils,
     relocation::{self, RelocatePayloadUtils, SignedRelocateDetailsUtils},
-    routing_api::core::{RESOURCE_PROOF_DATA_SIZE, RESOURCE_PROOF_DIFFICULTY},
     section::{
         test_utils::*, ElderCandidatesUtils, NodeStateUtils, SectionAuthorityProviderUtils,
         SectionKeyShare, SectionPeersUtils, SectionUtils, FIRST_SECTION_MIN_AGE, MIN_ADULT_AGE,
